+++ codex/serializers/opds_v1.py
@@ -43,13 +43,15 @@ class OPDSAcquisitionEntrySerializer(BrowserCardSerializer):
     language = CharField(read_only=True, source=UNIONFIX_PREFIX + "language")
 
     # ManyToMany
+    ## Categories
     characters = CharField(source=UNIONFIX_PREFIX + "characters")
-    genres = CharField(source=UNIONFIX_PREFIX + "genres")
-    locations = CharField(source=UNIONFIX_PREFIX + "locations")
-    series_groups = CharField(source=UNIONFIX_PREFIX + "series_groups")
+    # genres = CharField(source=UNIONFIX_PREFIX + "genres")
+    # locations = CharField(source=UNIONFIX_PREFIX + "locations")
+    # series_groups = CharField(source=UNIONFIX_PREFIX + "series_groups")
     story_arcs = CharField(source=UNIONFIX_PREFIX + "story_arcs")
-    tags = CharField(source=UNIONFIX_PREFIX + "tags")
-    teams = CharField(source=UNIONFIX_PREFIX + "teams")
+    # tags = CharField(source=UNIONFIX_PREFIX + "tags")
+    # teams = CharField(source=UNIONFIX_PREFIX + "teams")
+    ## Contributors
     authors = CharField(source=UNIONFIX_PREFIX + "authors")
     contributors = CharField(source=UNIONFIX_PREFIX + "contributors")
 
@@ -64,12 +66,12 @@ OPDS_FOLDERS_ORDERED_UNIONFIX_VALUES_MAP = get_serializer_values_map(
 
 OPDS_M2M_FIELDS = (
     "characters",
-    "genres",
-    "locations",
-    "series_groups",
+    # "genres",
+    # "locations",
+    # "series_groups",
     "story_arcs",
-    "tags",
-    "teams",
+    # "tags",
+    # "teams",
     "credits",
     "credits__role__name",
     "credits__person__name",